feat(Combined): extract components to files
Comfirmed, Vaccination,
